tools/SceneValidator/requirements.txt
pyyaml>=6.0
google-cloud-storage>=2.0.0
google-cloud-vision>=3.0.0
google-generativeai>=0.3.0
